Update CreateMap, GameObjectManager
Prompt:
"Update @MapType values to be paths relative to the  @backend/pkg/server/game_maps folder. Then update @CreateMap to call @CreateMapFromFile using the string value of the argument, and return the @BaseMap result. Update @GameObjectManager to maintain a pointer to a @Map and use it to keep track of map-related GameObjects. "